requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
sqlalchemy[asyncio]>=2.0.40,<3.0.0
asyncpg>=0.30.0,<1.0.0
aiosqlite>=0.21.0,<1.0.0
pydantic-settings>=2.9.0,<3.0.0
python-jose[cryptography]>=3.4.0,<4.0.0
passlib[argon2]>=1.7.4,<2.0.0
email-validator>=2.2.0,<3.0.0
python-multipart>=0.0.20,<1.0.0
